+++ src/stask/StreamingSink.cc
@@ -32,7 +32,7 @@ void StreamingSink::initialize() {
     mySenders = par("mySenders").stringValue();
     mySTaskCategory = par("mySTaskCategory").stringValue();
     lastCPUIndex = 0;
-    cpuCores = getAncestorPar("cores").longValue();
+    cpuCores = getAncestorPar("cores").intValue();
     if (cpuCores < 1) {
         throw new cRuntimeError("Number of CPU Cores is not set.");
     }
