+++ src/stask/StreamingSource.cc
@@ -36,7 +36,7 @@ void StreamingSource::initialize() {
 //    }
 //    calculateDelay();
     lastCPUIndex = 0;
-    cpuCores = getAncestorPar("cores").longValue();
+    cpuCores = getAncestorPar("cores").intValue();
     if (cpuCores < 1) {
         throw new cRuntimeError("Number of CPU Cores is not set.");
     }
